scrapers/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
schedule==1.2.0
Pillow==10.0.1
asyncio==3.4.3
aiohttp==3.9.1